pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kidney-hfm-eval"
version = "0.1.0"
description = "Evaluation framework for histopathology foundation models (HFMs) in kidney pathology"
readme = "readme.md"
requires-python = ">=3.9"
license = {text = "MIT"}

authors = [
  {name="Harishwar Reddy Kasireddy", email="harishwarreddy.k@ufl.edu"}
]

dependencies = [
    "numpy>=1.25",
    "pandas>=2.0",
    "scikit-learn>=1.2",
    "scipy>=1.10",
    "torch>=2.0",
    "torchvision>=0.15",
    "tqdm>=4.66",
    "matplotlib>=3.7",
    "seaborn>=0.13",
    "Pillow>=10.0"
]

[project.optional-dependencies]

wsi = [
    "tifffile",
    "tiffslide",
    "imageio",
    "scikit-image",
    "pyvips"
]

fms = [
    "timm>=1.0.0",
    "transformers>=4.36",
    "huggingface-hub>=0.20",
    "einops>=0.8",
    "fvcore",
    "iopath",
    "gigapath",
    "UNI",
    "vision-transformer"
]

[project.urls]
Homepage = "https://github.com/SarderLab/Feature-Extraction-from-FMs-and-Probing-.git"
